+++ clouddns/src/test/java/denominator/clouddns/CloudDNSZoneWriteCommandsLiveTest.java
@@ -1,0 +1,9 @@
+package denominator.clouddns;
+
+import denominator.Live.UseTestGraph;
+import denominator.ZoneWriteCommandsLiveTest;
+
+@UseTestGraph(CloudDNSTestGraph.class)
+public class CloudDNSZoneWriteCommandsLiveTest extends ZoneWriteCommandsLiveTest {
+
+}