+++ core/src/main/java/denominator/mock/MockZoneApi.java
@@ -11,12 +11,13 @@ import denominator.model.ResourceRecordSet;
 import denominator.model.Zone;
 import denominator.model.rdata.SOAData;
 
-import static denominator.common.Preconditions.checkArgument;
 import static denominator.common.Preconditions.checkState;
 import static denominator.common.Util.filter;
 import static denominator.model.ResourceRecordSets.nameAndTypeEqualTo;
 import static denominator.model.ResourceRecordSets.ns;
+import static denominator.model.ResourceRecordSets.soa;
 import static denominator.model.Zones.nameEqualTo;
+import static java.util.Arrays.asList;
 
 final class MockZoneApi implements denominator.ZoneApi {
 
@@ -32,22 +33,7 @@ final class MockZoneApi implements denominator.ZoneApi {
 
   MockZoneApi(Map<String, Collection<ResourceRecordSet<?>>> data) {
     this.data = data;
-    create("denominator.io.");
-  }
-
-  public void create(String name) {
-    checkArgument(!data.containsKey(name), "zone %s already exists", name);
-    Collection<ResourceRecordSet<?>> zone =
-        new ConcurrentSkipListSet<ResourceRecordSet<?>>(TO_STRING);
-    zone.add(ResourceRecordSet.builder()
-                 .type("SOA")
-                 .name(name)
-                 .ttl(86400)
-                 .add(SOAData.builder().mname("ns1." + name).rname("admin." + name)
-                          .serial(1).refresh(3600).retry(600).expire(604800).minimum(86400).build())
-                 .build());
-    zone.add(ns(name, 86400, "ns1." + name));
-    data.put(name, zone);
+    put(Zone.create("denominator.io.", "denominator.io.", 86400, "nil@denominator.io."));
   }
 
   @Override
@@ -83,5 +69,42 @@ final class MockZoneApi implements denominator.ZoneApi {
   @Override
   public Iterator<Zone> iterateByName(String name) {
     return filter(iterator(), nameEqualTo(name));
+  }
+
+  @Override
+  public String put(Zone zone) {
+    if (!data.containsKey(zone.name())) {
+      Collection<ResourceRecordSet<?>>
+          recordsInZone =
+          new ConcurrentSkipListSet<ResourceRecordSet<?>>(TO_STRING);
+      SOAData soaData = SOAData.builder().mname("ns1." + zone.name()).rname(zone.email())
+          .serial(1).refresh(3600).retry(600).expire(604800).minimum(86400).build();
+      recordsInZone.add(ResourceRecordSet.builder()
+                            .type("SOA")
+                            .name(zone.name())
+                            .ttl(zone.ttl())
+                            .add(soaData)
+                            .build());
+      recordsInZone.add(ns(zone.name(), zone.ttl(), asList("ns1." + zone.name())));
+      data.put(zone.name(), recordsInZone);
+      return zone.name();
+    }
+    for (Iterator<ResourceRecordSet<?>> i = data.get(zone.name()).iterator(); i.hasNext();) {
+      ResourceRecordSet<?> rrset = i.next();
+      if (rrset.type().equals("SOA")) {
+        SOAData soaData = (SOAData) rrset.records().get(0);
+        if (zone.email().equals(soaData.rname()) && zone.ttl() == rrset.ttl().intValue()) {
+          return zone.name();
+        }
+        i.remove();
+        data.get(zone.name()).add(soa(rrset, zone.email(), zone.ttl()));
+      }
+    }
+    return zone.name();
+  }
+
+  @Override
+  public void delete(String name) {
+    data.remove(name);
   }
 }