+++ dynect/src/main/java/denominator/dynect/DynECT.java
@@ -9,7 +9,6 @@ import java.util.List;
 import java.util.Map;
 
 import denominator.model.ResourceRecordSet;
-import denominator.model.Zone;
 import feign.Body;
 import feign.Headers;
 import feign.Param;
@@ -20,6 +19,13 @@ public interface DynECT {
 
   @RequestLine("GET /Zone")
   Data<List<String>> zones();;
+
+  @RequestLine("POST /Zone/{name}")
+  @Body("%7B\"ttl\":{ttl},\"rname\":\"{rname}\"%7D")
+  void createZone(@Param("name") String name, @Param("ttl") int ttl, @Param("rname") String rname);
+
+  @RequestLine("DELETE /Zone/{name}")
+  void deleteZone(@Param("name") String name);
 
   @RequestLine("PUT /Zone/{zone}")
   @Body("{\"publish\":true}")
@@ -54,6 +60,11 @@ public interface DynECT {
                             @Param("type") String type, @Param("ttl") int ttl,
                             @Param("rdata") Map<String, Object> rdata);
 
+  @RequestLine("PUT /SOARecord/{zone}/{zone}/{recordId}")
+  @Body("%7B\"ttl\":\"{ttl}\",\"rdata\":%7B\"rname\":\"{rname}\"%7D%7D")
+  void scheduleUpdateSOA(@Param("zone") String zone, @Param("recordId") long recordId,
+                         @Param("ttl") int ttl, @Param("rname") String rname);
+
   @RequestLine("DELETE /{recordId}")
   void scheduleDeleteRecord(@Param("recordId") String recordId);
 
@@ -65,21 +76,22 @@ public interface DynECT {
   /**
    * DynECT json includes an envelope called "data", which makes it difficult.
    */
-  static class Data<T> {
+  class Data<T> {
 
     T data;
   }
 
-  static class Record {
+  class Record {
 
     long id;
+    String serviceClass;
     String name;
     String type;
     int ttl;
     Map<String, Object> rdata = new LinkedHashMap<String, Object>();
   }
 
-  static class GeoService {
+  class GeoService {
 
     List<Node> nodes = new ArrayList<Node>();
     List<GeoRegionGroup> groups = new ArrayList<GeoRegionGroup>();