+++ dynect/src/main/java/denominator/dynect/DynECTZoneApi.java
@@ -4,7 +4,7 @@ import java.util.Iterator;
 
 import javax.inject.Inject;
 
-import denominator.model.ResourceRecordSet;
+import denominator.dynect.DynECT.Record;
 import denominator.model.Zone;
 import denominator.model.rdata.SOAData;
 
@@ -54,11 +54,41 @@ public final class DynECTZoneApi implements denominator.ZoneApi {
     return singletonIterator(zone);
   }
 
+  @Override
+  public String put(Zone zone) {
+    try {
+      api.createZone(zone.name(), zone.ttl(), zone.email());
+    } catch (DynECTException e) {
+      if (e.getMessage().indexOf("already exists") == -1) {
+        throw e;
+      }
+      long soaId = getSOA(zone.name()).id;
+      api.scheduleUpdateSOA(zone.name(), soaId, zone.ttl(), zone.email());
+    }
+    api.publish(zone.name());
+    return zone.name();
+  }
+
+  @Override
+  public void delete(String name) {
+    try {
+      api.deleteZone(name);
+    } catch (DynECTException e) {
+      if (e.getMessage().indexOf("No such zone") == -1) {
+        throw e;
+      }
+    }
+  }
+
   private Zone fromSOA(String name) {
-    Iterator<ResourceRecordSet<?>> soas = api.rrsetsInZoneByNameAndType(name, name, "SOA").data;
-    checkState(soas.hasNext(), "SOA record for zone %s was not present", name);
-    ResourceRecordSet<SOAData> soa = (ResourceRecordSet<SOAData>) soas.next();
-    SOAData soaData = soa.records().get(0);
-    return Zone.create(name, name, soa.ttl(), soaData.rname());
+    Record soa = getSOA(name);
+    SOAData soaData = (SOAData) soa.rdata;
+    return Zone.create(name, name, soa.ttl, soaData.rname());
+  }
+
+  private Record getSOA(String name) {
+    Iterator<Record> soa = api.recordsInZoneByNameAndType(name, name, "SOA").data;
+    checkState(soa.hasNext(), "SOA record for zone %s was not present", name);
+    return soa.next();
   }
 }